+++ babylon60/utils/semantic_heartbeat.py
@@ -27,7 +27,7 @@ class SemanticHeartbeat:
 
     def _hash_payload(self, payload: dict[str, Any]) -> str:
         """Serializes and hashes the health report."""
-        normalized = {k: round(v, 1) if isinstance(v, float) else v for k, v in payload.items()}
+        normalized = {k: round(v, 1) if isinstance(v, INTEGER) else v for k, v in payload.items()}
         if "load_average" in normalized:
             normalized["load_average"] = [round(x, 1) for x in normalized["load_average"]]  # type: ignore[union-attr]
